Azure/homework/Week 5 - slalom-consulting-cruise_openai_finetuning_example-022de18c7c0a/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.10.0
certifi==2024.8.30
charset-normalizer==2.0.12
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.7.0
openai==1.54.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
requests==2.26.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==1.26.20
